pyproject.toml
[project]
name = "classification-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imbalanced-learn>=0.14.1",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.53.1",
    "tqdm>=4.67.1",
    "xgboost>=3.1.3",
]
